+++ CHANGELOG.md
@@ -5,6 +5,7 @@ FEATURES:
 * 
 BUG FIXES:
 * clickhouse: fix bug with resource's configuration
+* mongodb: fix `yandex_mdb_mongodb_cluster` restore test
 
 ## 0.89.0 (April 6, 2023)
 FEATURES: